+++ examples/Context.tsx
@@ -1,5 +1,5 @@
 import { createContext, useContext, useRef } from "react";
-import { createScopedStore, scoped, ScopedContextValue } from "src";
+import { createScopedStore, ScopedContextValue, stateReq } from "src";
 
 type BearState = {
   bears: number;
@@ -13,7 +13,7 @@ type BearInitialData = {
 
 const createBearStore = createScopedStore<BearState, BearInitialData>()(
   (initialData) =>
-    scoped((set, get) => ({
+    stateReq((set, get) => ({
       bears: initialData.bears,
       increasePopulation: () =>
         set((state) => ({ ...state, bears: state.bears + 1 })),