+++ src/index.ts
@@ -5,4 +5,4 @@ export type {
 } from "./createScoped";
 export { createScopedStore } from "./createScoped";
 export { createScopedHook } from "./createScopedHook";
-export { scoped } from "./scoped";
+export { devtoolsReq, stateReq } from "./middleware";